commit de reparo
